+++ src/sif/nocost.cc
@@ -1,0 +1,217 @@
+#include "sif/nocost.h"
+#include "baldr/accessrestriction.h"
+#include "baldr/directededge.h"
+#include "baldr/graphconstants.h"
+#include "baldr/nodeinfo.h"
+#include "sif/costconstants.h"
+#include "sif/dynamiccost.h"
+
+#ifdef INLINE_TEST
+#include "test/test.h"
+#include "worker.h"
+#include <random>
+#endif
+
+using namespace valhalla::midgard;
+using namespace valhalla::baldr;
+
+namespace valhalla {
+namespace sif {
+
+/**
+ * Derived class providing dynamic edge costing for "direct" no-cost routes.
+ *
+ * Intended for use-cases where we dont care about mode of travel, this costing allows all edges.
+ */
+class NoCost : public DynamicCost {
+public:
+  /**
+   * Construct costing. Pass in cost type and options using protocol buffer(pbf).
+   * @param  costing specified costing type.
+   * @param  options pbf with request options.
+   */
+  NoCost(const Costing costing, const Options& options) : DynamicCost(options, TravelMode::kDrive) {
+  }
+
+  virtual ~NoCost() {
+  }
+
+  /**
+   * Get the access mode used by this costing method.
+   * @return  Returns access mode.
+   */
+  uint32_t access_mode() const {
+    return kAllAccess;
+  }
+
+  /**
+   * Checks if access is allowed for the provided directed edge.
+   * This is generally based on mode of travel and the access modes
+   * allowed on the edge. However, it can be extended to exclude access
+   * based on other parameters such as conditional restrictions and
+   * conditional access that can depend on time and travel mode.
+   * @param  edge           Pointer to a directed edge.
+   * @param  pred           Predecessor edge information.
+   * @param  tile           Current tile.
+   * @param  edgeid         GraphId of the directed edge.
+   * @param  current_time   Current time (seconds since epoch). A value of 0
+   *                        indicates the route is not time dependent.
+   * @param  tz_index       timezone index for the node
+   * @return Returns true if access is allowed, false if not.
+   */
+  virtual bool Allowed(const baldr::DirectedEdge* edge,
+                       const EdgeLabel& pred,
+                       const baldr::GraphTile*& tile,
+                       const baldr::GraphId& edgeid,
+                       const uint64_t current_time,
+                       const uint32_t tz_index,
+                       bool& has_time_restrictions) const {
+    return !edge->is_shortcut();
+  }
+
+  /**
+   * Checks if access is allowed for an edge on the reverse path
+   * (from destination towards origin). Both opposing edges (current and
+   * predecessor) are provided. The access check is generally based on mode
+   * of travel and the access modes allowed on the edge. However, it can be
+   * extended to exclude access based on other parameters such as conditional
+   * restrictions and conditional access that can depend on time and travel
+   * mode.
+   * @param  edge           Pointer to a directed edge.
+   * @param  pred           Predecessor edge information.
+   * @param  opp_edge       Pointer to the opposing directed edge.
+   * @param  tile           Current tile.
+   * @param  edgeid         GraphId of the opposing edge.
+   * @param  current_time   Current time (seconds since epoch). A value of 0
+   *                        indicates the route is not time dependent.
+   * @param  tz_index       timezone index for the node
+   * @return  Returns true if access is allowed, false if not.
+   */
+  virtual bool AllowedReverse(const baldr::DirectedEdge* edge,
+                              const EdgeLabel& pred,
+                              const baldr::DirectedEdge* opp_edge,
+                              const baldr::GraphTile*& tile,
+                              const baldr::GraphId& opp_edgeid,
+                              const uint64_t current_time,
+                              const uint32_t tz_index,
+                              bool& has_time_restrictions) const {
+    return !opp_edge->is_shortcut();
+  }
+
+  /**
+   * Checks if access is allowed for the provided node. Node access can
+   * be restricted if bollards or gates are present.
+   * @param  node  Pointer to node information.
+   * @return  Returns true if access is allowed, false if not.
+   */
+  virtual bool Allowed(const baldr::NodeInfo* node) const {
+    return true;
+  }
+
+  /**
+   * Only transit costings are valid for this method call, hence we throw
+   * @param edge
+   * @param departure
+   * @param curr_time
+   * @return
+   */
+  virtual Cost EdgeCost(const baldr::DirectedEdge* edge,
+                        const baldr::TransitDeparture* departure,
+                        const uint32_t curr_time) const {
+    throw std::runtime_error("NoCost::EdgeCost does not support transit edges");
+  }
+
+  /**
+   * Get the cost to traverse the specified directed edge. Cost includes
+   * the time (seconds) to traverse the edge.
+   * @param   edge    Pointer to a directed edge.
+   * @param   tile    Graph tile.
+   * @param   seconds Time of week in seconds.
+   * @return  Returns the cost and time (seconds)
+   */
+  virtual Cost EdgeCost(const baldr::DirectedEdge* edge,
+                        const baldr::GraphTile* tile,
+                        const uint32_t seconds) const {
+    return {static_cast<float>(edge->length()), static_cast<float>(edge->length())};
+  }
+
+  /**
+   * Returns the cost to make the transition from the predecessor edge.
+   * Defaults to 0. Costing models that wish to include edge transition
+   * costs (i.e., intersection/turn costs) must override this method.
+   * @param  edge  Directed edge (the to edge)
+   * @param  node  Node (intersection) where transition occurs.
+   * @param  pred  Predecessor edge information.
+   * @return  Returns the cost and time (seconds)
+   */
+  virtual Cost TransitionCost(const baldr::DirectedEdge* edge,
+                              const baldr::NodeInfo* node,
+                              const EdgeLabel& pred) const {
+    return {};
+  }
+
+  /**
+   * Returns the cost to make the transition from the predecessor edge
+   * when using a reverse search (from destination towards the origin).
+   * @param  idx   Directed edge local index
+   * @param  node  Node (intersection) where transition occurs.
+   * @param  pred  the opposing current edge in the reverse tree.
+   * @param  edge  the opposing predecessor in the reverse tree
+   * @return  Returns the cost and time (seconds)
+   */
+  virtual Cost TransitionCostReverse(const uint32_t idx,
+                                     const baldr::NodeInfo* node,
+                                     const baldr::DirectedEdge* pred,
+                                     const baldr::DirectedEdge* edge) const {
+    return {};
+  }
+
+  /**
+   * Get the cost factor for A* heuristics. This factor is multiplied
+   * with the distance to the destination to produce an estimate of the
+   * minimum cost to the destination. The A* heuristic must underestimate the
+   * cost to the destination. So a time based estimate based on speed should
+   * assume the maximum speed is used to the destination such that the time
+   * estimate is less than the least possible time along roads.
+   */
+  virtual float AStarCostFactor() const {
+    return 1.f;
+  }
+
+  /**
+   * Returns a function/functor to be used in location searching which will
+   * exclude and allow ranking results from the search by looking at each
+   * edges attribution and suitability for use as a location by the travel
+   * mode used by the costing method. Function/functor is also used to filter
+   * edges not usable / inaccessible by automobile.
+   */
+  virtual const EdgeFilter GetEdgeFilter() const {
+    // Throw back a lambda that checks the access for this type of costing
+    return [](const baldr::DirectedEdge* edge) {
+      return !(edge->is_shortcut() || edge->IsTransitLine());
+    };
+  }
+
+  /**
+   * Returns a function/functor to be used in location searching which will
+   * exclude results from the search by looking at each node's attribution
+   * @return Function/functor to be used in filtering out nodes
+   */
+  virtual const NodeFilter GetNodeFilter() const {
+    // throw back a lambda that checks the access for this type of costing
+    return [](const baldr::NodeInfo* node) { return false; };
+  }
+};
+
+void ParseNoCostOptions(const rapidjson::Document& doc,
+                        const std::string& costing_options_key,
+                        CostingOptions* pbf_costing_options) {
+  // this is probably not needed but its part of the contract for costing..
+}
+
+cost_ptr_t CreateNoCost(const Costing costing, const Options& options) {
+  return std::make_shared<NoCost>(costing, options);
+}
+
+} // namespace sif
+} // namespace valhalla